ASoC: rockchip: rk817-codec: Fix a large number of mclk counts being enabled
If it is not controlled, when the path is configured after
resume, the clock will be continuously turn on to increase
the clock count.

Therefore, we can count and control the switch of clock
separately according to the current status of playback and
capture, so as to avoid that mclk is accidentally turned off
when playback and capture exist at the same time, causing
the other stream to fail to work.

For example:
- before:
  mclk_sai0          2        2        0    11289600          0     0  50000
     mclk_sai0_out2io       1        1        0    11289600          0     0  50000
        mclk_sai0_to_io      16       16        0    11289600          0     0  50000

- after:
  mclk_sai0          2        2        0    11289600          0     0  50000
     mclk_sai0_out2io       1        1        0    11289600          0     0  50000
        mclk_sai0_to_io       2        2        0    11289600          0     0  50000

Change-Id: I78ec18c7ffc42f548e82357bcf20701aa057f15d
Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>